# Expose mining performance times in node status