--- a/ironfish/src/node.ts
+++ b/ironfish/src/node.ts
@@ -94,7 +94,7 @@ export class IronfishNode {
     this.chain = chain
     this.strategy = strategy
     this.metrics = metrics
-    this.miningManager = new MiningManager({ chain, memPool, node: this })
+    this.miningManager = new MiningManager({ chain, memPool, node: this, metrics })
     this.memPool = memPool
     this.workerPool = workerPool
     this.rpc = new RpcServer(this)