## Summary

This expose 3 new meters in our status related to mining when you pass --all to `ironfish status`

 - Time to construct a new block given transactions
 - Time to calculate transactions from the mem pool for the new block
 - Total time to generate a block template

```
> ironfish status --all

Version              0.1.45 @ src
Node                 STARTED
Node Name            JASON💪💎🚀
Block Graffiti       JASON💪💎🚀
Memory               Heap: 829.85 MiB -> 849.71 MiB / 4.00 GiB (20.2%), RSS: 2.68 GiB (16.8%), Free:
77.87 MiB (99.5%)
P2P Network          CONNECTED - In: 7.35 KB/s, Out: 1.88 KB/s, peers 3
Mining               STARTED - 1 miners, 0 mined, get txs: 0.0638ms, block: 30.5735ms, template:
253.2487ms
Mem Pool             Size: 0 tx, Bytes: 0 B
Syncer               IDLE - progress: 84.49%
Blockchain           NOT SYNCED @ HEAD 000000000005501b3df6494a0b34a5194971e7cefe288ae6c3e32dd2b9f3b8a9
 (136771)
Accounts             00000000000f730acf13ab7528473224b841ac851c502ef7685333410438298d (7803)
Telemetry            STARTED - 0 <- 1 pending
Workers              STARTED - 0 -> 1 / 6 - 0.05 jobs Δ, 408.52 jobs/s
```
Specifically this line
```
Mining               STARTED - 1 miners, 0 mined, get txs: 0.0638ms, block: 30.5735ms, template:
253.2487ms
```

## Testing Plan

## Breaking Change

Is this a breaking change? If yes, add notes below on why this is breaking and
what additional work is required, if any.

```
[ ] Yes
```